Moved `DrawingArea` set up into `Drawable` trait.
