+++ src/gui/hodograph/mod.rs
@@ -4,54 +4,105 @@ use gdk::EventMask;
 use gtk::prelude::*;
 use gtk::DrawingArea;
 
-use app::AppContextPointer;
+use gui::DrawingArgs;
+use gui::plot_context::{Drawable, GenericContext, HasGenericContext, PlotContextExt};
 
-pub mod hodo_context;
+use app::{config, AppContextPointer};
+use coords::{SDCoords, ScreenCoords, XYCoords};
 
 mod callbacks;
 mod drawing;
 
-pub fn set_up_hodograph_area(hodo_area: &DrawingArea, app_context: &AppContextPointer) {
-    hodo_area.set_hexpand(true);
-    hodo_area.set_vexpand(true);
+pub struct HodoContext {
+    generic: GenericContext,
+}
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_draw(move |_da, cr| callbacks::draw_hodo(cr, &ac));
+impl HodoContext {
+    pub fn new() -> Self {
+        HodoContext {
+            generic: GenericContext::new(),
+        }
+    }
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_scroll_event(move |da, ev| callbacks::scroll_event(da, ev, &ac));
+    pub fn convert_sd_to_xy(coords: SDCoords) -> XYCoords {
+        let radius = coords.speed / 2.0 / config::MAX_SPEED;
+        let angle = (270.0 - coords.dir).to_radians();
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_button_press_event(move |da, ev| callbacks::button_press_event(da, ev, &ac));
+        let x = radius * angle.cos() + 0.5;
+        let y = radius * angle.sin() + 0.5;
+        XYCoords { x, y }
+    }
 
-    let ac = Rc::clone(app_context);
-    hodo_area
-        .connect_button_release_event(move |da, ev| callbacks::button_release_event(da, ev, &ac));
+    pub fn convert_sd_to_screen(&self, coords: SDCoords) -> ScreenCoords {
+        let xy = HodoContext::convert_sd_to_xy(coords);
+        self.convert_xy_to_screen(xy)
+    }
+}
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_motion_notify_event(move |da, ev| callbacks::mouse_motion_event(da, ev, &ac));
+impl HasGenericContext for HodoContext {
+    fn get_generic_context(&self) -> &GenericContext {
+        &self.generic
+    }
+}
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_leave_notify_event(move |da, ev| callbacks::leave_event(da, ev, &ac));
+impl PlotContextExt for HodoContext {}
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_key_release_event(move |da, ev| callbacks::key_release_event(da, ev, &ac));
+impl Drawable for HodoContext {
+    fn set_up_drawing_area(da: &DrawingArea, acp: &AppContextPointer) {
+        use self::callbacks::*;
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_key_press_event(move |da, ev| callbacks::key_press_event(da, ev, &ac));
+        da.set_hexpand(true);
+        da.set_vexpand(true);
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_configure_event(move |da, ev| callbacks::configure_event(da, ev, &ac));
+        let ac = Rc::clone(acp);
+        da.connect_draw(move |_da, cr| draw_hodo(cr, &ac));
 
-    let ac = Rc::clone(app_context);
-    hodo_area.connect_size_allocate(move |da, ev| callbacks::size_allocate_event(da, ev, &ac));
+        let ac = Rc::clone(acp);
+        da.connect_scroll_event(move |da, ev| scroll_event(da, ev, &ac));
 
-    hodo_area.set_can_focus(true);
+        let ac = Rc::clone(acp);
+        da.connect_button_press_event(move |da, ev| button_press_event(da, ev, &ac));
 
-    hodo_area.add_events((EventMask::SCROLL_MASK | EventMask::BUTTON_PRESS_MASK
-        | EventMask::BUTTON_RELEASE_MASK
-        | EventMask::POINTER_MOTION_HINT_MASK
-        | EventMask::POINTER_MOTION_MASK | EventMask::LEAVE_NOTIFY_MASK
-        | EventMask::KEY_RELEASE_MASK | EventMask::KEY_PRESS_MASK)
-        .bits() as i32);
+        let ac = Rc::clone(acp);
+        da.connect_button_release_event(move |da, ev| button_release_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_motion_notify_event(move |da, ev| mouse_motion_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_leave_notify_event(move |da, ev| leave_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_key_release_event(move |da, ev| key_release_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_key_press_event(move |da, ev| key_press_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_configure_event(move |da, ev| configure_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_size_allocate(move |da, ev| size_allocate_event(da, ev, &ac));
+
+        da.set_can_focus(true);
+
+        da.add_events((EventMask::SCROLL_MASK | EventMask::BUTTON_PRESS_MASK
+            | EventMask::BUTTON_RELEASE_MASK
+            | EventMask::POINTER_MOTION_HINT_MASK
+            | EventMask::POINTER_MOTION_MASK | EventMask::LEAVE_NOTIFY_MASK
+            | EventMask::KEY_RELEASE_MASK | EventMask::KEY_PRESS_MASK)
+            .bits() as i32);
+    }
+
+    fn draw_background(&self, args: DrawingArgs) {
+        drawing::draw_background(args);
+    }
+
+    fn draw_data(&self, args: DrawingArgs) {
+        drawing::draw_data(args);
+    }
+
+    fn draw_overlays(&self, args: DrawingArgs) {
+        drawing::draw_overlays(args);
+    }
 }
