+++ src/gui/rh_omega/mod.rs
@@ -1,55 +1,203 @@
+use std::cell::Cell;
 use std::rc::Rc;
 
 use gdk::EventMask;
 use gtk::prelude::*;
 use gtk::DrawingArea;
 
-use app::AppContextPointer;
-
-pub mod rh_omega_context;
+use app::{config, AppContextPointer};
+use coords::{DeviceCoords, ScreenCoords, WPCoords, XYCoords};
+use gui::DrawingArgs;
+use gui::plot_context::{Drawable, GenericContext, HasGenericContext, PlotContext, PlotContextExt};
 
 mod drawing;
 mod callbacks;
 
-pub fn set_up_rh_omega_area(da: &DrawingArea, app_context: &AppContextPointer) {
-    da.set_hexpand(true);
-    da.set_vexpand(true);
+pub struct RHOmegaContext {
+    x_zoom: Cell<f64>,
+    generic: GenericContext,
+}
 
-    let ac = Rc::clone(app_context);
-    da.connect_draw(move |_da, cr| callbacks::draw_rh_omega(cr, &ac));
+impl RHOmegaContext {
+    pub fn new() -> Self {
+        RHOmegaContext {
+            x_zoom: Cell::new(1.0),
+            generic: GenericContext::new(),
+        }
+    }
 
-    let ac = Rc::clone(app_context);
-    da.connect_scroll_event(move |da, ev| callbacks::scroll_event(da, ev, &ac));
+    pub fn convert_wp_to_xy(coords: WPCoords) -> XYCoords {
+        use std::f64;
 
-    let ac = Rc::clone(app_context);
-    da.connect_button_press_event(move |da, ev| callbacks::button_press_event(da, ev, &ac));
+        let y = (f64::log10(config::MAXP) - f64::log10(coords.p))
+            / (f64::log10(config::MAXP) - f64::log10(config::MINP));
 
-    let ac = Rc::clone(app_context);
-    da.connect_button_release_event(move |da, ev| callbacks::button_release_event(da, ev, &ac));
+        // The + sign below looks weird, but is correct.
+        let x = (coords.w + config::MAX_ABS_W) / (2.0 * config::MAX_ABS_W);
 
-    let ac = Rc::clone(app_context);
-    da.connect_motion_notify_event(move |da, ev| callbacks::mouse_motion_event(da, ev, &ac));
+        XYCoords { x, y }
+    }
 
-    let ac = Rc::clone(app_context);
-    da.connect_leave_notify_event(move |da, ev| callbacks::leave_event(da, ev, &ac));
+    pub fn convert_xy_to_wp(coords: XYCoords) -> WPCoords {
+        use std::f64;
 
-    let ac = Rc::clone(app_context);
-    da.connect_key_release_event(move |da, ev| callbacks::key_release_event(da, ev, &ac));
+        let p = 10.0f64.powf(
+            -coords.y * (f64::log10(config::MAXP) - f64::log10(config::MINP))
+                + f64::log10(config::MAXP),
+        );
+        let w = coords.x * (2.0 * config::MAX_ABS_W) - config::MAX_ABS_W;
 
-    let ac = Rc::clone(app_context);
-    da.connect_key_press_event(move |da, ev| callbacks::key_press_event(da, ev, &ac));
+        WPCoords { w, p }
+    }
 
-    let ac = Rc::clone(app_context);
-    da.connect_configure_event(move |da, ev| callbacks::configure_event(da, ev, &ac));
+    pub fn convert_screen_to_wp(&self, coords: ScreenCoords) -> WPCoords {
+        let xy = self.convert_screen_to_xy(coords);
+        RHOmegaContext::convert_xy_to_wp(xy)
+    }
 
-    let ac = Rc::clone(app_context);
-    da.connect_size_allocate(move |da, ev| callbacks::size_allocate_event(da, ev, &ac));
+    pub fn convert_wp_to_screen(&self, coords: WPCoords) -> ScreenCoords {
+        let xy = RHOmegaContext::convert_wp_to_xy(coords);
+        self.convert_xy_to_screen(xy)
+    }
 
-    da.set_can_focus(true);
+    pub fn convert_device_to_wp(&self, coords: DeviceCoords) -> WPCoords {
+        let xy = self.convert_device_to_xy(coords);
+        Self::convert_xy_to_wp(xy)
+    }
 
-    da.add_events((EventMask::SCROLL_MASK | EventMask::BUTTON_PRESS_MASK
-        | EventMask::BUTTON_RELEASE_MASK | EventMask::POINTER_MOTION_HINT_MASK
-        | EventMask::POINTER_MOTION_MASK | EventMask::LEAVE_NOTIFY_MASK
-        | EventMask::KEY_RELEASE_MASK | EventMask::KEY_PRESS_MASK)
-        .bits() as i32);
+    pub fn set_translate_y(&self, new_translate: XYCoords) {
+        let mut translate = self.get_translate();
+        translate.y = new_translate.y;
+        self.set_translate(translate);
+    }
+}
+
+impl HasGenericContext for RHOmegaContext {
+    fn get_generic_context(&self) -> &GenericContext {
+        &self.generic
+    }
+}
+
+impl PlotContextExt for RHOmegaContext {
+    fn zoom_to_envelope(&self) {
+        let xy_envelope = self.get_xy_envelope();
+
+        let lower_left = xy_envelope.lower_left;
+        self.set_translate(lower_left);
+
+        let width = xy_envelope.upper_right.x - xy_envelope.lower_left.x;
+        let width_scale = 1.0 / width;
+
+        self.x_zoom.set(width_scale);
+    }
+
+    fn bound_view(&self) {
+        let device_rect = self.get_device_rect();
+
+        let bounds = DeviceCoords {
+            col: device_rect.width,
+            row: device_rect.height,
+        };
+        let lower_right = self.convert_device_to_xy(bounds);
+        let upper_left = self.convert_device_to_xy(device_rect.upper_left);
+        let height = upper_left.y - lower_right.y;
+
+        let mut translate = self.get_translate();
+        if height < 1.0 {
+            if translate.y < 0.0 {
+                translate.y = 0.0;
+            }
+            let max_y = 1.0 - height;
+            if translate.y > max_y {
+                translate.y = max_y;
+            }
+        } else {
+            translate.y = -(height - 1.0) / 2.0;
+        }
+        self.set_translate(translate);
+    }
+
+    fn convert_xy_to_screen(&self, coords: XYCoords) -> ScreenCoords {
+        // Apply translation first
+        let x = coords.x - self.get_translate().x;
+        let y = coords.y - self.get_translate().y;
+
+        // Apply scaling
+        let x = x * self.x_zoom.get();
+        let y = y * self.get_zoom_factor();
+
+        ScreenCoords { x, y }
+    }
+
+    fn convert_screen_to_xy(&self, coords: ScreenCoords) -> XYCoords {
+        // Unapply scaling first
+        let x = coords.x / self.x_zoom.get();
+        let y = coords.y / self.get_zoom_factor();
+
+        // Unapply translation
+        let x = x + self.get_translate().x;
+        let y = y + self.get_translate().y;
+
+        XYCoords { x, y }
+    }
+}
+
+impl Drawable for RHOmegaContext {
+    fn set_up_drawing_area(da: &DrawingArea, acp: &AppContextPointer) {
+        use self::callbacks::*;
+
+        da.set_hexpand(true);
+        da.set_vexpand(true);
+
+        let ac = Rc::clone(acp);
+        da.connect_draw(move |_da, cr| draw_rh_omega(cr, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_scroll_event(move |da, ev| scroll_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_button_press_event(move |da, ev| button_press_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_button_release_event(move |da, ev| button_release_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_motion_notify_event(move |da, ev| mouse_motion_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_leave_notify_event(move |da, ev| leave_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_key_release_event(move |da, ev| key_release_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_key_press_event(move |da, ev| key_press_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_configure_event(move |da, ev| configure_event(da, ev, &ac));
+
+        let ac = Rc::clone(acp);
+        da.connect_size_allocate(move |da, ev| size_allocate_event(da, ev, &ac));
+
+        da.set_can_focus(true);
+
+        da.add_events((EventMask::SCROLL_MASK | EventMask::BUTTON_PRESS_MASK
+            | EventMask::BUTTON_RELEASE_MASK
+            | EventMask::POINTER_MOTION_HINT_MASK
+            | EventMask::POINTER_MOTION_MASK | EventMask::LEAVE_NOTIFY_MASK
+            | EventMask::KEY_RELEASE_MASK | EventMask::KEY_PRESS_MASK)
+            .bits() as i32);
+    }
+
+    fn draw_background(&self, args: DrawingArgs) {
+        drawing::draw_background(args);
+    }
+
+    fn draw_data(&self, args: DrawingArgs) {
+        drawing::draw_data(args);
+    }
+
+    fn draw_overlays(&self, args: DrawingArgs) {
+        drawing::draw_overlays(args);
+    }
 }
